Refactored "status" field of question model
The "status" field used to describe whether a question was 'drafted','open' or
'closed'. This information has now become obsolete as the status of a
question can be seen from the "end_date" field.
Is the "end_date" null, then it is a draft.
Is the "end_date" > now, then it is open.
Is the "end_date" <= now, then it is closed.
In order to maintain the convenient querying functionality of django a
"QuestionManager" was added. In addition the question model now has a
function "is_closed()".