Add permission to see exam / votes during board meeting.
The 'examsite' users in the remote olympiad should not see the
exam, votes, downloads, etc. during the Board Meeting, only later
when the final PDFs are released for printing.

This commit introduces a permission 'can_see_boardmeeting' that
should be given to delegations, staff, authors, etc. that _can_
see the exam and related things during the board meeting.