Add a `Dockerfile` which contains a multi-stage build for the different docker images used in ExamTools development, or in the (experimental) Kubernetes deployment.
Add two compose files mirroring the current `a-dev` and `b-dev`
development setups.

Instructions for how to build / launch the Docker compose setup, 
and how to attach a debugger to a server, are added to the main
repository `README.md`.

From CI, the images to be used in the Kubernetes deployment are
pushed, but the current test / image build steps are not (yet)
replaced.

Follow-up tasks:
- Add SFTP
- Use the multistage build for the main pipeline images
- Improve docker caching in CI

Unrelated changes:
Turn off Docker caching in the pipelines and always execute all 
jobs, to get the pipeline back to a passing state.